Add PubSub APIs.
This commit adds the new API's for the `pubsub`
and `et` (Encore Testing) packages.

It was auto generated using `encr.dev/tools/publicapigen`
from the [Encore runtime](https://github.com/encoredev/encore/tree/main/runtime)
and is the first time the tool has been used, so all other changes in this commit
are the result of the formatting of the public api generator.